+++ lib/binance_client/models/market.rb
@@ -27,15 +27,20 @@ module BinanceClient
       is_margin_trading_allowed: :margin_trading_allowed?,
     }.freeze
 
-    attr_reader :raw_hash
+    attr_accessor :raw_hash
+    attr_writer *METHODS
 
-    def initialize(raw_hash:)
-      @raw_hash = raw_hash
+    def initialize(**kwargs)
+      kwargs.each do |key, value|
+        self.send("#{key}=", value)
+      end
     end
 
     METHODS.each do |method_name|
       define_method method_name do
-        raw_hash[method_name.to_s.camelcase(:lower)]
+        memoize_results "@#{method_name}" do
+          raw_hash[method_name.to_s.camelcase(:lower)]
+        end
       end
     end