+++ lib/binance_client/responses/coins_info_response.rb
@@ -1,4 +1,11 @@
 module BinanceClient
   class CoinsInfoResponse < BaseResponse
+
+    def coins
+      @coins ||= body.map do |coin_hash|
+        Coin.new(raw_hash: coin_hash)
+      end
+    end
+
   end
 end