+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- `Coin#network_list` is initialized properly
+
 ## [3.0.0]
 ### Changed
 - Rename `#get_all` to `#coins_info` and give access to info via the response's `#coins` method