+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0]
+### Added
+- Add optional `network` parameter when getting deposit address of a specific network
+- Use keyword arguments when calling `sub_account_deposit_address`
+
 ## [3.0.1]
 ### Fixed
 - `Coin#network_list` is initialized properly